extract cached_name
